server/py3/requirements.txt
black>=24.0.0
isort>=5.13.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
alembic>=1.13.0
greenlet>=3.0.0
psycopg2-binary>=2.9.0

fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
python-dotenv

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
aiosqlite>=0.19.0